==> R/06_about_index.Rmd <==
---
output: html_document
---

## Motivation

When we travel somewhere new, checking the weather is simple. But checking the water quality is not.  

This webpsite is a "weather app for water quality" that makes public water quality data accessible and understandable, enabling Californians to check public drinking water quality in their hometown, place of work, school, and anywhere they may travel.  

***  

## Background

Every year, nearly 40 million Californians turn on the tap to receive water. For the great majority of Californians, this water is supplied by a public water system. The state government of California requires public water systems to independently compile water quality data for their service area, and send out an annual water quality report to their customers, called a consumer confidence report (CCR). These reports suffer from a lack of consistency across agencies, sometimes fail to effectively communicate water quality data, and are only updated once per year. Near-real-time, standardized, and easy to understand CCRs would improve the public's understanding of the water they buy and consume.    


***

## Open data and code

The data used to build these reports are gathered and curated by the California State Water Resources Control Board, and available to the public online. You can find water quality data [here](https://www.waterboards.ca.gov/water_issues/programs/hr2w/) and [here](https://www.waterboards.ca.gov/drinking_water/certlic/drinkingwater/documents/edtlibrary/), and water system spatial boundaries [here](https://data.ca.gov/dataset/drinking-water-water-system-service-area-boundaries). Another online version of CCRs, and a statewide water system search can be found [here](https://sdwis.waterboards.ca.gov/PDWW/JSP/WaterSystemDetail.jsp?tinwsys_is_number=4424&tinwsys_st_code=CA&wsnumber=CA4510005).  

The code underlying this webpage can be found [here on Github](https://github.com/caccr/caccr.github.io). The code that creates the reports and webpages can be found [here on Github](https://github.com/caccr/cawdc_2019). For questions related to the codebase, contact the principal developer of this project, [Rich Pauloo](richpauloo.github.io).   

***  

## Frequently asked questions

[Click here to find answers to frequently asked questions about this website](https://caccr.github.io/faq/).  

The answer to [frequently asked questions about water quality in the state of California may be found in this PDF](https://www.waterboards.ca.gov/water_issues/programs/hr2w/docs/general/faqs_6-05-2018.pdf).  



***

## Acknowledgements

We acknowledge the [California State Water Resources Control Board](https://www.waterboards.ca.gov/) for the original idea to pursue CCRs as a project, and the [West Big Data Hub](https://westbigdatahub.org/) for convening the datathon where the first lines of code of this project were written. We thank water quality experts in the California state government who helped interpret state water quality databases, and the [Internet of Water](https://internetofwater.org/) for helping review early versions of this website. And last, but not least, we thank [Lindsay Poirier](http://lindsaypoirier.com/), Avery Kruger, and Andrew Essin for their webscraping, data visualization, and water quality research assistance.  


<br> 

<br>

<br> 

<br>
